+++ src/main/java/chess/domain/piece/move/PawnCatchMove.java
@@ -1,4 +1,4 @@
-package chess.domain.position.move;
+package chess.domain.piece.move;
 
 public final class PawnCatchMove implements PieceMove {
