+++ src/main/java/chess/domain/position/PiecesPosition.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.position;
 
 import chess.domain.piece.Bishop;
 import chess.domain.piece.Camp;
@@ -8,9 +8,6 @@ import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Queen;
 import chess.domain.piece.Rook;
-import chess.domain.position.File;
-import chess.domain.position.Position;
-import chess.domain.position.Rank;
 import java.util.LinkedHashMap;
 import java.util.Map;
