+++ src/test/java/chess/domain/piece/KingTest.java
@@ -5,9 +5,9 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import chess.domain.position.File;
 import chess.domain.position.Position;
 import chess.domain.position.Rank;
-import chess.domain.position.move.BlockingMove;
-import chess.domain.position.move.InvalidMove;
-import chess.domain.position.move.PieceMove;
+import chess.domain.piece.move.BlockingMove;
+import chess.domain.piece.move.InvalidMove;
+import chess.domain.piece.move.PieceMove;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
