+++ src/test/java/chess/domain/piece/PawnTest.java
@@ -5,10 +5,10 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import chess.domain.position.File;
 import chess.domain.position.Position;
 import chess.domain.position.Rank;
-import chess.domain.position.move.InvalidMove;
-import chess.domain.position.move.PawnCatchMove;
-import chess.domain.position.move.PawnForwardMove;
-import chess.domain.position.move.PieceMove;
+import chess.domain.piece.move.InvalidMove;
+import chess.domain.piece.move.PawnCatchMove;
+import chess.domain.piece.move.PawnForwardMove;
+import chess.domain.piece.move.PieceMove;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
